moved code into src
